chore: reorganize files and imports
