+++ bricks/r13n/hooks/exceptions/arb_missing_region_tag_exception.dart
@@ -1,0 +1,9 @@
+import '../hooks.dart';
+
+class ArbMissingRegionTagException extends R13nException {
+  const ArbMissingRegionTagException()
+      : super(
+    message:
+    'Missing region tag in arb file, make sure to include @@region',
+  );
+}
