+++ bricks/r13n/hooks/exceptions/r13n_yaml_not_found_exception.dart
@@ -1,0 +1,8 @@
+import '../hooks.dart';
+
+class R13nYamlNotFoundException extends R13nException {
+  R13nYamlNotFoundException()
+      : super(
+    message: 'No r13n.yaml found.',
+  );
+}
