+++ bricks/r13n/hooks/hooks.dart
@@ -1,5 +1,6 @@
 export 'arb/arb.dart';
+export 'exceptions/exceptions.dart';
 export 'post_gen.dart';
+export 'process_overrides/process_overrides.dart';
 export 'r13n_configuration.dart';
-export 'r13n_exception.dart';
 export 'r13n_version_compatibility.dart';
