+++ bricks/r13n/hooks/r13n_configuration.dart
@@ -32,9 +32,3 @@ class R13nConfiguration {
   final String templateArbFile;
 }
 
-class R13nYamlNotFoundException extends R13nException {
-  R13nYamlNotFoundException()
-      : super(
-          message: 'No r13n.yaml found.',
-        );
-}
