+++ bricks/r13n/hooks/test/src/exceptions/arb_missing_region_tag_exception_test.dart
@@ -2,7 +2,7 @@
 
 import 'package:test/test.dart';
 
-import '../../../pre_gen.dart';
+import '../../../hooks.dart';
 
 void main() {
   group('ArbMissingRegionTagException', () {
